# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: tea121-lite backend
  version: 0.1.0
paths:
...
  /api/v1/runs/{run_id}/summary:
    get:
      tags:
      - results
      summary: Summary
      operationId: summary_api_v1_runs__run_id__summary_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummaryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/alarms:
    get:
      tags:
      - results
      summary: Alarms
      operationId: alarms_api_v1_runs__run_id__alarms_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: detector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Detector Id
      - name: rule_pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Pack Id
      - name: cwe_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      - name: violation_kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Violation Kind
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: function
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Function
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AlarmOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/alarms/{alarm_id}:
    get:
      tags:
      - results
      summary: Alarm
      operationId: alarm_api_v1_runs__run_id__alarms__alarm_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: alarm_id
        in: path
        required: true
        schema:
          type: string
          title: Alarm Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlarmOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/diagnostics:
    get:
      tags:
      - results
      summary: Diagnostics
      operationId: diagnostics_api_v1_runs__run_id__diagnostics_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_DiagnosticOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/cfg:
    get:
      tags:
      - results
      summary: Cfg
      operationId: cfg_api_v1_runs__run_id__cfg_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: function
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Function
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/states:
    get:
      tags:
      - results
      summary: States
      operationId: states_api_v1_runs__run_id__states_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: function
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Function
      - name: block_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Block Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/ir:
    get:
      tags:
      - results
      summary: Ir
      operationId: ir_api_v1_runs__run_id__ir_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: function
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Function
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/trace:
    get:
      tags:
      - results
      summary: Trace
      operationId: trace_api_v1_runs__run_id__trace_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: function
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Function
      - name: block_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Block Id
      - name: after
        in: query
        required: false
        schema:
          type: integer
          default: -1
          title: After
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 500
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AlarmOut:
      properties:
        id:
          type: string
          title: Id
        run_id:
          type: string
          title: Run Id
        alarm_key:
          type: string
          title: Alarm Key
        detector_id:
          type: string
          title: Detector Id
        detector_version:
          type: string
          title: Detector Version
        rule_pack_id:
          type: string
          title: Rule Pack Id
        rule_pack_version:
          type: string
          title: Rule Pack Version
        cwe_id:
          type: string
          title: Cwe Id
        family:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
        violation_kind:
          type: string
          title: Violation Kind
        severity:
          type: string
          enum:
          - definite
          - possible
          title: Severity
        function_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Function Name
        block_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Block Id
        instruction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Instruction Id
        instruction_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Instruction Text
        source_file_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File Id
        source_line:
          anyOf:
          - type: integer
          - type: 'null'
          title: Source Line
        memory_object_id:
          type: string
          title: Memory Object Id
        object_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Object Name
        object_size:
          $ref: '#/components/schemas/Interval'
        offset:
          $ref: '#/components/schemas/Interval'
        access_size_bytes:
          type: integer
          minimum: 0.0
          title: Access Size Bytes
        safe_condition:
          type: string
          title: Safe Condition
        message:
          type: string
          title: Message
        reason:
          items:
            type: string
          type: array
          title: Reason
        evidence:
          additionalProperties: true
          type: object
          title: Evidence
      type: object
      required:
      - id
      - run_id
      - alarm_key
      - detector_id
      - detector_version
      - rule_pack_id
      - rule_pack_version
      - cwe_id
      - violation_kind
      - severity
      - memory_object_id
      - object_size
      - offset
      - access_size_bytes
      - safe_condition
      - message
      - reason
      title: AlarmOut
...
    DiagnosticOut:
      properties:
        id:
          type: string
          title: Id
        run_id:
          type: string
          title: Run Id
        diagnostic_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Diagnostic Id
        code:
          type: string
          title: Code
        severity:
          type: string
          enum:
          - unknown_effect
          - unsupported
          - error
          title: Severity
        message:
          type: string
          title: Message
        location:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Location
        impact:
          type: string
          title: Impact
        function_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Function Name
        block_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Block Id
        instruction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Instruction Id
        source_file_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File Id
        source_line:
          anyOf:
          - type: integer
          - type: 'null'
          title: Source Line
      type: object
      required:
      - id
      - run_id
      - code
      - severity
      - message
      - impact
      title: DiagnosticOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Interval:
      properties:
        lower:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lower
        upper:
          anyOf:
          - type: integer
          - type: 'null'
          title: Upper
        is_bottom:
          type: boolean
          title: Is Bottom
          default: false
      type: object
      title: Interval
...
    Page_AlarmOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AlarmOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[AlarmOut]
...
    Page_DiagnosticOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DiagnosticOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[DiagnosticOut]
...
    SummaryOut:
      properties:
        run_id:
          type: string
          title: Run Id
        status:
          type: string
          title: Status
        detector_id:
          type: string
          title: Detector Id
        rule_pack_id:
          type: string
          title: Rule Pack Id
        alarm_count:
          type: integer
          title: Alarm Count
        definite_count:
          type: integer
          title: Definite Count
        possible_count:
          type: integer
          title: Possible Count
        diagnostic_count:
          type: integer
          title: Diagnostic Count
        unsupported_count:
          type: integer
          title: Unsupported Count
        error_count:
          type: integer
          title: Error Count
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        functions_analyzed:
          type: integer
          title: Functions Analyzed
      type: object
      required:
      - run_id
      - status
      - detector_id
      - rule_pack_id
      - alarm_count
      - definite_count
      - possible_count
      - diagnostic_count
      - unsupported_count
      - error_count
      - duration_ms
      - functions_analyzed
      title: SummaryOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError